Changed marker tap behaviour

+++ app/src/main/java/com/example/sayyaf/homecare/mapping/MapsActivity.java
@@ -74,7 +74,7 @@ public class MapsActivity extends AppCompatActivity implements OnMapReadyCallbac
     private static final String FINE_LOCATION = Manifest.permission.ACCESS_FINE_LOCATION;
     private static final String COARSE_LOCATION = Manifest.permission.ACCESS_COARSE_LOCATION;
     private static final int LOCATION_REQUEST_CODE = 1;
-    private static final float STREET_ZOOM = 17;
+    private static final int STREET_ZOOM = 17;
     private static final LatLngBounds LAT_LNG_BOUNDS =
             new LatLngBounds(new LatLng(-90, -180), new LatLng(90, 180));
